pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "narrative-os"
version = "0.1.0"
description = "可编程叙事操作系统 (Narrative OS) — 面向长篇小说的叙事操作系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # 核心数据校验
    "pydantic>=2.7",
    # 图结构 (PlotGraph)
    "networkx>=3.3",
    # 向量数据库 (MemorySystem)
    "chromadb>=0.5",
    # LLM 调用 + LangGraph 调度
    "langchain>=0.2",
    "langchain-openai>=0.1",
    "langchain-anthropic>=0.1",
    "langchain-community>=0.2",
    "langgraph>=0.1",
    # Web API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    # CLI
    "typer>=0.12",
    "rich>=13.7",
    # 配置 & 序列化
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    # Embedding (本地 fallback)
    "sentence-transformers>=3.0",
    # 进程管理（DevServer 端口检测与子进程清理）
    "psutil>=5.9",
    # 结构化数据库
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
viz = [
    # 图可视化 (PlotGraph -> Graphviz) — 需要系统安装 Graphviz
    "pygraphviz>=1.13",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",        # FastAPI test client
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
narrative = "narrative_os.interface.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["narrative_os"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
